.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
mysql-connector-python
torch
torchvision
pillow
timm
